controlserver/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
playwright>=1.41.0
playwright-stealth>=1.0.6
pyperclip>=1.8.0
pyyaml>=6.0
pydantic>=2.5.0
